requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
PyJWT
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
numpy
scikit-learn